Refactor utils into the existing modules
