+++ docs/reference.md
@@ -1,3 +1,2 @@
 ::: kachel.server
 ::: kachel.cache
-::: kachel.utils
